tools/llm-eval/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-eval"
version = "0.1.0"
description = "Reusable LLM evaluation harness for Story App research spikes"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "anthropic>=0.30",
    "google-genai>=1.0",
    "mistralai>=1.0",
    "jsonschema>=4.0",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "Pillow>=10.0",
    "scipy>=1.10",
    "lxml>=5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
